// src/types/settings.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SalesPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from './role';

export interface Settings {
  id: number;
  shopName: string;
  ownerLabel: string;
  attendantLabel: string;
  activeRole: Role;
  lowStockThreshold: number;
}