// src/types/productUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SalesPulse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  categoryId?: number;
  barcode?: string;
  /** @minimum 0 */
  sellingPriceCents?: number;
  /** @minimum 0 */
  costPriceCents?: number;
  /** @minimum 0 */
  lowStockThreshold?: number;
}