// src/types/customerLedgerEntry.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SalesPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerLedgerEntryType } from './customerLedgerEntryType';

export interface CustomerLedgerEntry {
  id: number;
  customerId: number;
  type: CustomerLedgerEntryType;
  amountCents: number;
  /** @nullable */
  note: string | null;
  /** @nullable */
  saleId: number | null;
  createdAt: Date;
}